requirements.txt
asyncpraw==7.8.1
pydantic-ai-slim[openai]==0.0.24
asyncpg==0.30.0
fastapi==0.115.8
logfire[asyncpg]==3.6.1
python-dotenv==1.0.1
html2text==2024.2.26
google-api-python-client==2.161.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
chardet==5.2.0
pillow==10.4.0